+++ src/core/services/twig/functions/getNestedValue.spec.ts
