+++ src/core/services/twig/functions/getNestedValue.ts
