+++ src/core/services/twig/twigService.ts
@@ -1,29 +1,7 @@
 import Twig from 'twig';
-import type { TwigConfiguration } from '../../../types';
-import { getNestedValue } from '../utils/getNestedValue';
-
-const sort = (a: any, b: any): number => {
-    if (!a) {
-        return 0;
-    }
-    if (a < b) {
-        return -1;
-    }
-    if (a > b) {
-        return 1;
-    }
-    return 0;
-};
-
-const sortBy = (value: any, key: any): any => {
-    if (!Array.isArray(value)) {
-        return value;
-    }
-
-    value.sort((a, b) => sort(a[key as string], b[key as string]));
-
-    return value;
-};
+import type { TwigConfiguration } from '../../../../types';
+import { getNestedValue } from './functions/getNestedValue';
+import { sortBy } from './filters/sortBy';
 
 class TwigService {
     public static render(
